+++ packages/imperative/__tests__/src/packages/profiles/__integration__/ProfileCommandExample.integration.test.ts
@@ -129,7 +129,7 @@ describe("Imperative should allow CLI implementations to configure their own pro
     }
 
     // eslint-disable-next-line jest/expect-expect, jest/no-disabled-tests
-    it.skip("We should be able to issue a command and have a profile be automatically loaded", function () {
+    it("We should be able to issue a command and have a profile be automatically loaded", function () {
         // return issueCommand.bind(this, ["profile"], true)();
     });
 });