+++ packages/imperative/src/cmd/src/profiles/CommandProfiles.ts
@@ -62,40 +62,6 @@ export class CommandProfiles {
     }
 
     /**
-     * Gets the first (or by name) profile in the map - automatically throws an exception (unless disabled)
-     * @deprecated Load profile properties from `IHandlerParameters.arguments` property instead.
-     * @template T - The expected profile mapping to be returned
-     * @param {string} type - The profile type
-     * @param {string} [name=""] - The name of the profile to retrieve
-     * @param {boolean} [failNotFound=true] - Automatically throws an imperative exception if not profiles are not
-     * found - this is provided as convince for the handlers (will fail your command if not found) - This would
-     * normally be the result of a command configuration problem.
-     * @returns {T} - The first profile in the map (or the one located by name)
-     * @memberof CommandProfiles
-     */
-    public get<T extends IProfile>(type: string, failNotFound = true, name = ""): T {
-        let profile: IProfile;
-        // If a profile is returned for the type, then we'll check if a profile of a specific name was requseted
-        // if not, just return the first profile found (first loaded)
-        if (this.map.get(type) != null) {
-            if (name != null && name.trim().length > 0) {
-                for (const prof of this.map.get(type)) {
-
-                    if (prof.name === name) {
-                        profile = prof;
-                        break;
-                    }
-                }
-            } else {
-                profile = this.map.get(type)[0];
-            }
-        } else if (failNotFound) {
-            this.fail(type);
-        }
-        return profile as T;
-    }
-
-    /**
      * Gets all profiles for the type specified,
      * @deprecated
      * @template T - extends IProfile